# [TraceQL] Fix subtly incorrect handling of second pass conditions